Test searchByRank function
